backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.0.0
google-genai
opencv-python
ffmpeg-python
playwright
decord
python-dotenv